+++ HISTORY.rst
@@ -4,6 +4,9 @@
 ++++++++++++++++
 
 - Removing dependence on ``acor`` extension.
+- Added arguments to ``PTSampler`` function.
+- Added automatic load-balancing for MPI runs.
+- Added custom load-balancing for MPI and multiprocessing.
 
 
 2.0.0 (2013-11-17)